rename cvrp folder
