+++ cvrp_metaheuristics/algorithm/greedy_algorithm.py
@@ -1,9 +1,9 @@
 import sys
 
-from cvpr_metaheuristics.algorithm.result import Result
-from cvpr_metaheuristics.problem.cvrp import cost
+from cvrp_metaheuristics.algorithm.result import Result
+from cvrp_metaheuristics.problem.cvrp import cost
 
-from cvpr_metaheuristics.algorithm.algorithm import Algorithm
+from cvrp_metaheuristics.algorithm.algorithm import Algorithm
 
 
 class GreedyAlgorithm(Algorithm):
