+++ cvrp_metaheuristics/utils/enums.py
