+++ setup.py
@@ -1,3 +1,3 @@
 from setuptools import setup, find_packages
 
-setup(name='cvpr_metaheuristics', version='0.0', packages=find_packages())
+setup(name='cvrp_metaheuristics', version='0.0', packages=find_packages())
